# optimize time zone lookups