--- a/crates/jiff-static/lib.rs
+++ b/crates/jiff-static/lib.rs
@@ -46,7 +46,7 @@ use quote::quote;
 
 use self::shared::{
     util::array_str::Abbreviation, PosixDay, PosixDayTime, PosixDst,
-    PosixOffset, PosixRule, PosixTime, PosixTimeZone, TzifFixed,
+    PosixOffset, PosixRule, PosixTime, PosixTimeZone, TzifDateTime, TzifFixed,
     TzifIndicator, TzifLocalTimeType, TzifOwned, TzifTransitionInfo,
     TzifTransitionKind, TzifTransitionsOwned,
 };
@@ -259,14 +259,10 @@ impl TzifTransitionsOwned {
             ref civil_ends,
             ref infos,
         } = *self;
-        let civil_starts: Vec<_> = civil_starts
-            .iter()
-            .map(|(y, mo, d, h, m, s)| quote!((#y, #mo, #d, #h, #m, #s)))
-            .collect();
-        let civil_ends: Vec<_> = civil_ends
-            .iter()
-            .map(|(y, mo, d, h, m, s)| quote!((#y, #mo, #d, #h, #m, #s)))
-            .collect();
+        let civil_starts: Vec<_> =
+            civil_starts.iter().map(TzifDateTime::quote).collect();
+        let civil_ends: Vec<_> =
+            civil_ends.iter().map(TzifDateTime::quote).collect();
         let infos: Vec<_> =
             infos.iter().map(TzifTransitionInfo::quote).collect();
         quote! {
@@ -343,6 +339,27 @@ impl TzifTransitionKind {
             TzifTransitionKind::Fold => quote! {
                 jiff::shared::TzifTransitionKind::Fold
             },
+        }
+    }
+}
+
+impl TzifDateTime {
+    fn quote(&self) -> proc_macro2::TokenStream {
+        let year = self.year();
+        let month = self.month();
+        let day = self.day();
+        let hour = self.hour();
+        let minute = self.minute();
+        let second = self.second();
+        quote! {
+            jiff::shared::TzifDateTime::new(
+                #year,
+                #month,
+                #day,
+                #hour,
+                #minute,
+                #second,
+            )
         }
     }
 }