--- a/crates/jiff-static/shared/mod.rs
+++ b/crates/jiff-static/shared/mod.rs
@@ -237,16 +237,6 @@ pub struct TzifTransitionInfo {
     pub kind: TzifTransitionKind,
 }
 
-/// The representation we use to represent a civil datetime.
-///
-/// The tuple is years, months, days, hours, minutes and seconds.
-///
-/// We don't use `shared::util::itime::IDateTime` here because we
-/// specifically do not need to represent fractional seconds. This
-/// lets us easily represent what we need in 8 bytes instead of
-/// the 12 bytes used by `IDateTime`.
-pub type TzifDateTime = (i16, i8, i8, i8, i8, i8);
-
 /// The kind of a transition.
 ///
 /// This is used when trying to determine the offset for a local datetime. It
@@ -352,6 +342,73 @@ pub enum TzifTransitionKind {
     /// local time. Also has an entry in `civil_ends` for when the fold ends
     /// (exclusive) in local time.
     Fold,
+}
+
+/// The representation we use to represent a civil datetime.
+///
+/// We don't use `shared::util::itime::IDateTime` here because we specifically
+/// do not need to represent fractional seconds. This lets us easily represent
+/// what we need in 8 bytes instead of the 12 bytes used by `IDateTime`.
+///
+/// Moreover, we explicitly set the alignment here to force the compiler to
+/// optimize comparisons down to `i64.cmp(&i64)`. This is especially useful
+/// since we do a binary search on `&[TzifDateTime]` when doing a TZ lookup for
+/// a civil datetime.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd, Ord)]
+#[repr(align(8))]
+pub struct TzifDateTime {
+    year: i16,
+    month: i8,
+    day: i8,
+    hour: i8,
+    minute: i8,
+    second: i8,
+}
+
+impl TzifDateTime {
+    pub const ZERO: TzifDateTime = TzifDateTime {
+        year: 0,
+        month: 0,
+        day: 0,
+        hour: 0,
+        minute: 0,
+        second: 0,
+    };
+
+    pub const fn new(
+        year: i16,
+        month: i8,
+        day: i8,
+        hour: i8,
+        minute: i8,
+        second: i8,
+    ) -> TzifDateTime {
+        TzifDateTime { year, month, day, hour, minute, second }
+    }
+
+    pub const fn year(self) -> i16 {
+        self.year
+    }
+
+    pub const fn month(self) -> i8 {
+        self.month
+    }
+
+    pub const fn day(self) -> i8 {
+        self.day
+    }
+
+    pub const fn hour(self) -> i8 {
+        self.hour
+    }
+
+    pub const fn minute(self) -> i8 {
+        self.minute
+    }
+
+    pub const fn second(self) -> i8 {
+        self.second
+    }
 }
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]