--- a/crates/jiff-static/shared/tzif.rs
+++ b/crates/jiff-static/shared/tzif.rs
@@ -5,7 +5,7 @@
 use alloc::{string::String, vec};
 
 use super::{
-    util::{Byte, Bytes},
+    util::escape::{Byte, Bytes},
     PosixTimeZone, TzifFixed, TzifIndicator, TzifLocalTimeType, TzifOwned,
     TzifTransition,
 };