--- a/crates/jiff-static/shared/util/array_str.rs
+++ b/crates/jiff-static/shared/util/array_str.rs
@@ -139,6 +139,12 @@ impl<const N: usize> core::fmt::Write for ArrayStr<N> {
     }
 }
 
+impl<const N: usize> AsRef<str> for ArrayStr<N> {
+    fn as_ref(&self) -> &str {
+        self.as_str()
+    }
+}
+
 /// A self-imposed limit on the size of a time zone abbreviation, in bytes.
 ///
 /// POSIX says this: