… CHANGELOG.md
… COMPARE.md
… Cargo.toml
… bench/Cargo.toml
… bench/src/datetime.rs
… bench/src/tz.rs
… bench/src/zoned.rs
… crates/jiff-static/Cargo.toml
… crates/jiff-static/shared/util/error.rs
… crates/jiff-static/shared/util/escape.rs
… crates/jiff-static/shared/util/mod.rs
… crates/jiff-static/shared/util/utf8.rs
… src/civil/date.rs
… src/civil/datetime.rs
… src/civil/time.rs
… src/civil/weekday.rs
… src/error.rs
… src/fmt/strtime/mod.rs
… src/fmt/temporal/parser.rs
… src/fmt/temporal/printer.rs
… src/lib.rs
… src/shared/mod.rs
… src/shared/posix.rs
… src/shared/tzif.rs
… src/shared/util/array_str.rs
… src/shared/util/error.rs
… src/shared/util/escape.rs
… src/shared/util/itime.rs
… src/shared/util/mod.rs
… src/shared/util/utf8.rs