This PR introduces a number of internal changes that
overall make time zone lookups faster across the board:

```
group                                          base                                   opt4
-----                                          ----                                   ----
tz/posix_datetime_to_offset/jiff               2.52     82.9±1.14ns        ? ?/sec    1.00     32.9±0.28ns        ? ?/sec
tz/posix_timestamp_to_offset/jiff              2.11     47.7±0.26ns        ? ?/sec    1.00     22.6±0.15ns        ? ?/sec
tz/tzif_bundled_datetime_to_offset/jiff        11.52   104.9±0.45ns        ? ?/sec    1.00      9.1±0.06ns        ? ?/sec
tz/tzif_bundled_timestamp_to_offset/jiff       7.78     48.4±0.39ns        ? ?/sec    1.00      6.2±0.05ns        ? ?/sec
tz/tzif_future_datetime_to_offset/jiff         2.81    104.9±0.67ns        ? ?/sec    1.00     37.4±0.67ns        ? ?/sec
tz/tzif_future_timestamp_to_offset/jiff        2.28     48.3±0.23ns        ? ?/sec    1.00     21.2±0.17ns        ? ?/sec
tz/tzif_historical_datetime_to_offset/jiff     2.68     23.4±0.13ns        ? ?/sec    1.00      8.7±0.08ns        ? ?/sec
tz/tzif_historical_timestamp_to_offset/jiff    1.42      8.5±0.06ns        ? ?/sec    1.00      6.0±0.05ns        ? ?/sec
```

The big change here has to do with automatically "fattening" TZif
data to reduce the likelihood of a TZ lookup going through a POSIX
time zone. In general, POSIX time zone lookups are slower than TZif
lookups when the input time matches a range defined by an explicit
transition. But whether Jiff uses the slower POSIX time zone path or
not is largely data dependent on the TZif data. Some TZif data (from
`/usr/share/zoneinfo`) is "slim," where only the transitions that are
absolutely necessary for historical reasons are present. Other TZif
data (also from `/usr/share/zoneinfo`) can be "fat," where there are
many more explicit transitions that extend into the future.

Whether TZif data is fat or not generally depends on how your platform
generates the TZif data in `/usr/share/zoneinfo`. If they're optimizing
for size, they'll generate slim data. If they're optimizing for
compatibility, they'll generate fat data (because I guess some
applications still can't handle POSIX time zones???).

Exacerbating this issue is that `jiff-tzdb` uses "slim" data in order
to optimize for binary size since it's embedded into compiled binaries.

This became very apparent in issues like #271, where one might want to
switch to Jiff's new `jiff::tz::get!(..)` macro if you only ever care
about one time zone. Prior to this PR, there was a very noticeable
performance regression if you had originally been using fat TZif data
from `/usr/share/zoneinfo`, and this was largely because of going
through the POSIX time zone path.

This PR resolves this discrepancy by "auto-fattening" TZif data with
additional transitions in a way that should match what "fat" TZif
data looks like. The downside is that this will result in about 20%
more heap memory (for time zones loaded from disk at runtime) or 20%
more space in the binary (for time zones embedded into the binary via
`jiff::tz::get!(..)`). In absolute terms, TZif data uses very little
memory, so this additional usage shouldn't be a burden. For users
that really care, they can disable the new `tz-fat` feature (which
is enabled by default) in order to prioritize less memory usage over
runtime performance.

While that on its own lead to a nice speed-up, profiling revealed
that the bottleneck was now in binary search. So a second class of
optimization in this PR is making binary search for TZ lookups faster.
This basically boiled down to making the sequence we search over itself
smaller, thereby improving cache locality.

Optimizing binary search was done by first observing that TZ lookups
for timestamps only searched timestamps, and TZ lookups for local
datetimes only searched local datetimes. So instead of a single
sequence with "fat" elements containing the timestamp and the local
datetimes (in addition to some other data), we switch to a "array" or
column based storage. I also optimized the representation of local
datetimes in TZif data _specifically_ to make it smaller (since we
don't need fractional seconds) and to make comparisons themselves much
faster (by bit-packing a civil datetime into a single `i64`).

Looking back to #271, I made this program inspired by the original
MRE:

```rust
use std::time::Instant;

use jiff::Timestamp;

fn main() -> anyhow::Result<()> {
    let timestamps_chrono: Vec<i64> =
        (0..50_000_000).map(|offset| 1620000000000000000 + offset).collect();

    let timestamps_jiff: Vec<Timestamp> = (0..50_000_000)
        .map(|offset| {
            Timestamp::from_nanosecond(1620000000000000000 + offset).unwrap()
        })
        .collect();

    chrono_static_bundle(&timestamps_chrono);
    jiff_static_bundle(&timestamps_jiff);
    jiff_static_zoneinfo(&timestamps_jiff);
    jiff_runtime_zoneinfo(&timestamps_jiff);

    Ok(())
}

#[allow(dead_code)]
#[inline(never)]
fn chrono_static_bundle(timestamps: &[i64]) {
    use chrono::TimeZone;

    let start = Instant::now();
    timestamps.iter().for_each(|&timestamp| {
        chrono_tz::America::New_York.timestamp_nanos(timestamp);
    });
    println!("Chrono took: {:?}", start.elapsed());
}

#[allow(dead_code)]
#[inline(never)]
fn jiff_static_bundle(timestamps: &[Timestamp]) {
    static TZ: jiff::tz::TimeZone = jiff::tz::get!("America/New_York");
    let start = Instant::now();
    timestamps.iter().for_each(|timestamp| {
        timestamp.to_zoned(TZ.clone());
    });
    println!("Jiff[static-bundled] took: {:?}", start.elapsed());
}

#[allow(dead_code)]
#[inline(never)]
fn jiff_static_zoneinfo(timestamps: &[Timestamp]) {
    static TZ: jiff::tz::TimeZone =
        jiff::tz::include!("/usr/share/zoneinfo/America/New_York");
    let start = Instant::now();
    timestamps.iter().for_each(|timestamp| {
        timestamp.to_zoned(TZ.clone());
    });
    println!("Jiff[static-zoneinfo] took: {:?}", start.elapsed());
}

#[allow(dead_code)]
#[inline(never)]
fn jiff_runtime_zoneinfo(timestamps: &[Timestamp]) {
    let tz = jiff::tz::TimeZone::get("America/New_York").unwrap();
    let start = Instant::now();
    timestamps.iter().for_each(|timestamp| {
        timestamp.to_zoned(tz.clone());
    });
    println!("Jiff[runtime-zoneinfo] took: {:?}", start.elapsed());
}
```

Running this program in release mode on `master` gives this output:

```
$ ./target/release/i271-regression-from-chrono
Chrono took: 638.125112ms
Jiff[static-bundled] took: 3.053803415s
Jiff[static-zoneinfo] took: 792.087182ms
Jiff[runtime-zoneinfo] took: 958.111947ms
```

With this PR, we now get:

```
$ ./target/release/i271-regression-from-chrono
Chrono took: 643.60309ms
Jiff[static-bundled] took: 613.653771ms
Jiff[static-zoneinfo] took: 610.679669ms
Jiff[runtime-zoneinfo] took: 733.377423ms
```

The `static-zoneinfo` and `runtime-zoneinfo` Jiff benchmarks were using
"fat" TZif data, which explains why they were already decently
fast. (Their speed-ups here come from improving cache locality of
binary search.) But the `static-bundled` benchmark was using "slim"
data and thus going through POSIX time zones for TZ lookups. That
gets a huge speed-up.

This benchmark for static time zones is now faster than Chrono, while
the runtime lookup is slightly slower. The static benchmark is actually
an apples-to-apples comparison, since this program uses `chrono-tz`,
which exclusively provides static time zones (i.e., compiled into your
binary). The Jiff runtime benchmark is only slower here because of the
`Arc` clone (of the runtime loaded `jiff::tz::TimeZone`) needed for the
workflow described in #271. Even so, it is nearly as fast as Chrono.

Closes #271